add NumericSymbolButton
