+++ src/components/NumericInput/index.tsx
@@ -30,6 +30,7 @@
 import NumericInputComponent from './NumericInput';
 import NumericInputContainer from './primitives/NumericInputContainer';
 import NumericSymbol from './primitives/NumericSymbol';
+import NumericSymbolButton from './primitives/NumericSymbolButton';
 import NumericTextInput from './primitives/NumericTextInput';
 
 const NumericInput = Object.assign(NumericInputComponent, {
@@ -38,6 +39,9 @@ const NumericInput = Object.assign(NumericInputComponent, {
 
     /** Renders its children as the symbol (currency or unit) displayed beside the number. */
     Symbol: NumericSymbol,
+
+    /** Renders a pressable symbol (currency or unit) selector. */
+    SymbolButton: NumericSymbolButton,
 
     /** Renders the centered, full-size amount layout with legacy empty-area refocus behavior. */
     Container: NumericInputContainer,
